flaskr/requirements.txt
flask
openai
google-generativeai
pinecone-client
langchain
python-dotenv
pypdf2
